+++ package.json
@@ -12,6 +12,7 @@
     "react-icons": "^4.3.1",
     "react-loading": "^2.0.3",
     "react-redux": "^8.0.1",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
     "redux-thunk": "^2.4.1",
     "web-vitals": "^2.1.4"
