+++ src/pages/editToDo.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-function NewItem() {
+function EditItem() {
   return (
     <div class="editor mx-auto w-10/12 flex flex-col text-gray-800 border border-gray-300 p-4 shadow-lg max-w-2xl mt-10">
       <h1>Name</h1>
@@ -18,12 +18,12 @@ function NewItem() {
       ></textarea>
 
       <div class="buttons flex justify-end pt-10">
-        <div class="btn border border-blue-500  p-1 px-4 font-semibold cursor-pointer text-gray-200 ml-2 bg-blue-500 hover:bg-blue-700">
-          Submit
+        <div class="btn border border-blue-500  p-1 px-4 font-semibold cursor-pointer text-gray-200 ml-2 bg-orange-500   hover:bg-orange-700">
+          Edit
         </div>
       </div>
     </div>
   );
 }
 
-export default NewItem;
+export default EditItem;
