+++ src/pages/index.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import NavHome from '../component/navHome.jsx';
+import ListToDo from '../component/listToDo.jsx';
+
+function HomePage() {
+  return (
+    <div>
+      <NavHome />
+      <ListToDo />
+    </div>
+  );
+}
+
+export default HomePage;
